auth_service/requirements.txt
Django==5.2.6
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
